Move few functions from query to utils
